HSEARCH-4173 Separate the test profiles for Elasticsearch 6.7 and 6.8
