+++ pom.xml
@@ -219,6 +219,7 @@
         <version.org.elasticsearch.latest-7.6>7.6.2</version.org.elasticsearch.latest-7.6>
         <version.org.elasticsearch.latest-7.2>7.2.1</version.org.elasticsearch.latest-7.2>
         <version.org.elasticsearch.latest-6.8>6.8.9</version.org.elasticsearch.latest-6.8>
+        <version.org.elasticsearch.latest-6.7>6.7.2</version.org.elasticsearch.latest-6.7>
         <version.org.elasticsearch.latest-6.6>6.6.2</version.org.elasticsearch.latest-6.6>
         <version.org.elasticsearch.latest-6.3>6.3.2</version.org.elasticsearch.latest-6.3>
         <version.org.elasticsearch.latest-6.2>6.2.4</version.org.elasticsearch.latest-6.2>
